pyproject.toml
[project]
name = "sowtee-backend"
version = "0.1.0"
description = "SOWTEE - Agentic AAC Communication Bridge Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-generativeai>=0.3.0",
    "chromadb>=0.4.22",
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "groq>=0.4.0",
    "lingodotdev>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
